query-tester/requirements.txt
psycopg2-binary
neo4j
elasticsearch
